move thread-based parmap implementation into new module
