+++ lib/spack/spack/util/multithread.py
@@ -1,0 +1,38 @@
+##############################################################################
+# Copyright (c) 2013-2018, Lawrence Livermore National Security, LLC.
+# Produced at the Lawrence Livermore National Laboratory.
+#
+# This file is part of Spack.
+# Created by Todd Gamblin, tgamblin@llnl.gov, All rights reserved.
+# LLNL-CODE-647188
+#
+# For details, see https://github.com/spack/spack
+# Please also see the NOTICE and LICENSE files for our notice and the LGPL.
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License (as
+# published by the Free Software Foundation) version 2.1, February 1999.
+#
+# This program is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the IMPLIED WARRANTY OF
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the terms and
+# conditions of the GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public
+# License along with this program; if not, write to the Free Software
+# Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+##############################################################################
+from multiprocessing.dummy import Pool
+
+
+MAX_THREADS_PER_MAP = 5
+
+
+def parmap(f, X):
+    # Note that the dummy version of Pool uses threads rather than processes.
+    # Using a global thread pool shared by all parmap calls would cause
+    # hangs if the function provided to parmap also invokes parmap.
+    if not X:
+        return list()
+    pool = Pool(min(len(X), MAX_THREADS_PER_MAP))
+    return list(pool.map(f, X))
